elasticacheTools/requirements.txt
boto3==1.20.37
botocore==1.23.37
environs==9.4.0
jmespath==0.10.0
marshmallow==3.14.1
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.8
